+++ frontend/src/main.jsx
@@ -4,10 +4,14 @@ import { AuthProviderWrapper } from './auth/AuthProviderWrapper'
 import App from './App.jsx'
 import './index.css'
 
+import { BrowserRouter } from 'react-router-dom'
+
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <AuthProviderWrapper>
-      <App />
-    </AuthProviderWrapper>
+    <BrowserRouter>
+      <AuthProviderWrapper>
+        <App />
+      </AuthProviderWrapper>
+    </BrowserRouter>
   </React.StrictMode>,
 )